Conflicts:
	WorkOrder/.idea/misc.xml
	WorkOrder/app/src/main/AndroidManifest.xml
	WorkOrder/app/src/main/java/com/xiangxun/workorder/bean/ChildData.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/bean/GroupData.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/EquipmentMenuAcitvity.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/MainActivity.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/MaintenanceActivity.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/adapter/DetailOrderImageAdapter.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/adapter/ExpandableListViewAdapter.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/adapter/MainDetailImageAdapter.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/biz/DetailOrderFragmentListener.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/fragment/DetailOrderFragment.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/main/ShowImageViewActivity.java
	WorkOrder/app/src/main/java/com/xiangxun/workorder/ui/presenter/DetailOrderFragmentPresenter.java
	WorkOrder/app/src/main/res/layout/item_child_expandablelistview.xml
	WorkOrder/app/src/main/res/layout/item_group_expandablelistview.xml

+++ WorkOrder/app/src/main/java/com/xiangxun/workorder/widget/camera/PhotoPop.java
@@ -33,9 +33,13 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
+/**
+ * @TODO：拍照，相册选择弹出框。
+ */
 public class PhotoPop extends PopupWindow implements OnClickListener {
 
     private String rota = "ROTATE.png";
+    private int qualityCompress = AppEnum.WIDTH.getLen() * AppEnum.HEIGHT.getLen() / 10;//设备屏幕像素的1/10
 
     public PhotoPop(Context context) {
         super();
@@ -135,7 +139,7 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
      */
     private void capPhoto() {
         // 照相
-        capUri = AppEnum.ROOT + System.currentTimeMillis() + ".png";
+        capUri = AppEnum.ROOT + rota;
         AppEnum.capUri = capUri;
         File destDir = new File(AppEnum.ROOT);
         if (!destDir.exists()) {
@@ -192,10 +196,15 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
         DLog.i("PopStringActivityResult");
         switch (tag) {
             case AppEnum.REQUESTCODE_CAP:
+                String ppp = getImagePathForCAP(capUri);
+                DLog.i(ppp);
+                if (ppp == null) {
+                    break;
+                }
                 // 照相机程序返回的
-                String url = copyFile(getImagePathForCAP(capUri), AppEnum.IMAGE + System.currentTimeMillis() + ".png");
+                String url = copyFile(ppp, AppEnum.IMAGE + System.currentTimeMillis() + ".png");
                 //需要删除以前文件。
-                File file = new File(getImagePathForCAP(capUri));
+                File file = new File(ppp);
                 if (file.exists()) {
                     file.delete();
                 }
@@ -205,7 +214,15 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
                 // 照片的原始资源地址
                 photoUri = data.getData();
                 DLog.i(getImagePath(photoUri));
-                String cam = copyFile(getImagePath(photoUri), AppEnum.IMAGE + System.currentTimeMillis() + ".png");
+                if (getImagePath(photoUri) == null) {
+                    break;
+                }
+                String cam = copyFile(getImagePathForCAP(getImagePath(photoUri)), AppEnum.IMAGE + System.currentTimeMillis() + ".png");
+                //需要删除以前文件。
+                File idk = new File(getImagePathForCAP(getImagePath(photoUri)));
+                if (idk.exists()) {
+                    idk.delete();
+                }
                 return cam;
             default:
                 break;
@@ -233,7 +250,6 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
                 int length;
                 while ((byteread = inStream.read(buffer)) != -1) {
                     bytesum += byteread; //字节数 文件大小
-                    System.out.println(bytesum);
                     fs.write(buffer, 0, byteread);
                 }
                 inStream.close();
@@ -257,13 +273,9 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
         }
         // 由于相机拍照照片过大。故而必须进行压缩。
         int degree = getBitmapDegree(capUri);
-        if (degree == 0) {
-            return capUri;
-        } else {
-            // 先压缩图片。然后旋转图片。最后保存图片。
-            rotateBitmapByDegree(compressImageCap(capUri), degree);
-            return AppEnum.ROOT + rota;
-        }
+        // 先压缩图片。然后旋转图片。最后保存图片。
+        rotateBitmapByDegree(compressImageCap(capUri), degree);
+        return AppEnum.ROOT + rota;
     }
 
     /**
@@ -280,7 +292,8 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
             if (file_size > 320) {
                 bit = decodeSampledBitmapFromFile(capUri);
                 // 获取图片大小的比对关系。是100KB的多少。
-                int quality = 1024 * 100 / file_size;
+                int quality = qualityCompress / file_size * 10;
+                DLog.i(getClass().getSimpleName(), quality + "----" + file_size);
                 return compressBitmap(bit, quality);
             } else {
                 try {
@@ -334,6 +347,7 @@ public class PhotoPop extends PopupWindow implements OnClickListener {
                 targetwidth = picwidth / inSampleSize;
             }
         }
+        DLog.i(getClass().getSimpleName(), "压缩比例" + inSampleSize);
         return inSampleSize;
     }
